requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-spectacular
asgiref==3.9.1
Django==5.2.5
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
boto3
